cash-918 another method of getting options to render and work. Renaming a pillToggle file to something moore specific.
